+++ src/components/WidgetForm/SimpleSearch/validators.js
